requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg[binary]==3.2.3
django-cors-headers==4.6.0
django-redis==5.4.0
channels==4.2.0
channels-redis==4.2.1
django-storages[boto3]==1.14.4
whitenoise==6.8.2
requests==2.32.3
Pillow==11.0.0
gunicorn==23.0.0
